requirements.txt
v20
tpqoax
mlflow
pandas
scikit-learn
numpy
pyyaml
matplotlib
tensorflow
keras
streamlit
dagshub
dvc
dvc-s3
python-dotenv